refactor: Move ramp range slider into a subcomponent
